# schemas/openapi.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "NeverLauncher API",
    "version": "1.0.0",
    "description": "Канонический production API NeverLauncher 0.19.5. Исторические маршруты /api/v2–/api/v5 удалены и намеренно не входят в контракт."
  },
  "servers": [
    {
      "url": "/",
      "description": "Текущий Backend NeverLauncher"
    }
  ],
...
  "paths": {
...
    "/api/v1/server-bridge/servers": {
      "get": {
        "operationId": "get_server_bridge_servers",
        "tags": [
          "bridge"
        ],
        "summary": "GET /api/v1/server-bridge/servers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/server-bridge/servers/{serverId}/control/{commandId}": {
      "get": {
        "operationId": "get_server_bridge_servers_by_serverId_control_by_commandId",
        "tags": [
          "bridge"
        ],
        "summary": "GET /api/v1/server-bridge/servers/{serverId}/control/{commandId}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerBridgeControlCommandResponseV3"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "commandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "NodeId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-NeverLauncher-Node-Id",
        "description": "ServerBridge node id included in the Ed25519 canonical request."
      },
      "NodeKeyFingerprint": {
        "type": "apiKey",
        "in": "header",
        "name": "X-NeverLauncher-Node-Key-Fingerprint",
        "description": "Lowercase SHA-256 fingerprint of the registered raw Ed25519 public key."
      },
      "NodeTimestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "X-NeverLauncher-Node-Timestamp",
        "description": "Unix timestamp in seconds; must be within the ServerBridge signing window."
      },
      "NodeNonce": {
        "type": "apiKey",
        "in": "header",
        "name": "X-NeverLauncher-Node-Nonce",
        "description": "Fresh unpadded base64url nonce, consumed once by PostgreSQL replay protection."
      },
      "NodeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-NeverLauncher-Node-Signature",
        "description": "Unpadded base64url Ed25519 signature over NeverLauncher-ServerBridge-Node-v1 canonical request."
      },
      "BootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-NeverLauncher-Bootstrap-Token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required, node signature invalid, or signed request outside the accepted timestamp window",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}